Remove windows specific warnings

+++ src/common.cpp
@@ -9,14 +9,14 @@ device_info_t getDeviceInfo(cl::Device &d)
 {
     device_info_t devInfo;
     
-    devInfo.numCUs = d.getInfo<CL_DEVICE_MAX_COMPUTE_UNITS>();
-    devInfo.maxWGSize = d.getInfo<CL_DEVICE_MAX_WORK_GROUP_SIZE>();
+    devInfo.numCUs = (uint)d.getInfo<CL_DEVICE_MAX_COMPUTE_UNITS>();
+    devInfo.maxWGSize = (uint)d.getInfo<CL_DEVICE_MAX_WORK_GROUP_SIZE>();
     // Limiting max work-group size to 256
     #define MAX_WG_SIZE 256
     devInfo.maxWGSize = MIN(devInfo.maxWGSize, MAX_WG_SIZE);
     
-    devInfo.maxAllocSize = (unsigned int)d.getInfo<CL_DEVICE_MAX_MEM_ALLOC_SIZE>();
-    devInfo.maxGlobalSize = (unsigned int)d.getInfo<CL_DEVICE_GLOBAL_MEM_SIZE>();
+    devInfo.maxAllocSize = (uint)d.getInfo<CL_DEVICE_MAX_MEM_ALLOC_SIZE>();
+    devInfo.maxGlobalSize = (uint)d.getInfo<CL_DEVICE_GLOBAL_MEM_SIZE>();
     devInfo.doubleSupported = false;
     
     std::string extns = d.getInfo<CL_DEVICE_EXTENSIONS>();
